# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"SignUpReqBody": {
				"properties": {
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"name",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AddNoteReqBody": {
				"properties": {
					"title": {
						"type": "string"
					},
					"body": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"body"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EditNoteReqBody": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"body": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "notes-app",
		"version": "0.0.0",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/auth/signup": {
			"post": {
				"operationId": "SignUp",
				"responses": {
					"200": {
						"description": "Success"
					}
				},
				"description": "Sign up a user.",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SignUpReqBody"
							}
						}
					}
				}
			}
		},
		"/note/getnotes": {
			"get": {
				"operationId": "GetNotes",
				"responses": {
					"200": {
						"description": "Success"
					}
				},
				"description": "Add a note for the user.",
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/note/addnote": {
			"post": {
				"operationId": "AddNote",
				"responses": {
					"200": {
						"description": "Success"
					}
				},
				"description": "Add a note for the user.",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddNoteReqBody"
							}
						}
					}
				}
			}
		},
		"/note/editnote": {
			"post": {
				"operationId": "EditNote",
				"responses": {
					"200": {
						"description": "Success"
					}
				},
				"description": "Add a note for the user.",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EditNoteReqBody"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}